// src/types/karyawan.ts
export interface Karyawan {
    id:  number;
    name: string;
    email: string;
    role: string;
    password?: string;
    namaLengkap?: string;
    namaPerusahaan?: string;
    namaPanggilan?: string;
    jk?: string;
    employeeId?: string;
    tempatLahir?: string;
    noTelp?: string;
    golDarah?: string;
    statusMenikah?: string;
    agama?: string;
    jabatan?: string;
    statusKerja?: string;
    tanggalMasuk?: string;
    tanggalLahir?: string;
    pendidikan?: string;
    fakultas?: string;
    jurusan?: string;
    universitas?: string;
    swastaNegeri?: string;
    ipk?: string;
    domisili?: string;
    alamatDomisili?: string;
    alamatKtp?: string;
    noKtp?: string;
    noKk?: string;
    bpjsTk?: string;
    bpjsKes?: string;
    statusPtkp?: string;
    kategoriPph?: string;
    npwp?: string;
    kartuAsuransi?: string;
    memberNumber?: string;
    noPolis?: string;
    emailPribadi?: string;
    emailPerusahaan?: string;
    kontakDaruratNama?: string;
    kontakDaruratHubungan?: string;
    kontakDaruratHp?: string;
    ig?: string;
    linkedin?: string;
    bank?: string;
    noRekening?: string;
    namaRekening?: string;
    [key: string]: string | number | undefined;
  }